attendances

+++ Dawem.Domain/Entities/Schedules/ShiftWorkingTime.cs
@@ -18,7 +18,7 @@ namespace Dawem.Domain.Entities.Schedules
         public AmPm TimePeriod { get; set; }
         public TimeOnly CheckInTime { get; set; }
         public TimeOnly CheckOutTime { get; set; }
-        public double AllowedMinutes { get; set; }
+        public int AllowedMinutes { get; set; }
