+++ Dawem.Models/Response/Schedules/ShiftWorkingTimes/GetShiftWorkingTimeByIdResponseDTO.cs
@@ -9,7 +9,7 @@ namespace Dawem.Models.Response.Schedules.ShiftWorkingTimes
         public string Name { get; set; }
         public TimeOnly CheckInTime { get; set; }
         public TimeOnly CheckOutTime { get; set; }
-        public double AllowedMinutes { get; set; }
+        public int AllowedMinutes { get; set; }
         public bool IsActive { get; set; }
         public AmPm TimePeriod { get; set; }
