+++ Dawem.Models/Response/Schedules/ShiftWorkingTimes/GetShiftWorkingTimeResponseModelDTO.cs
@@ -10,7 +10,7 @@ namespace Dawem.Models.Response.Schedules.ShiftWorkingTimes
         public string Name { get; set; }
         public TimeOnly CheckInTime { get; set; }
         public TimeOnly CheckOutTime { get; set; }
-        public double AllowedMinutes { get; set; }
+        public int AllowedMinutes { get; set; }
         public bool IsActive { get; set; }
     }
 }
